refactor(app): move everything under the `regular` feature module
