+++ src/lib/features/drafts/timeline/regular/index.svelte
@@ -1,18 +1,15 @@
 <script lang="ts">
-  import ChevronDownIcon from '@lucide/svelte/icons/chevron-down';
   import FlaskConicalIcon from '@lucide/svelte/icons/flask-conical';
   import GraduationCapIcon from '@lucide/svelte/icons/graduation-cap';
   import PaperclipIcon from '@lucide/svelte/icons/paperclip';
 
-  import * as DropdownMenu from '$lib/components/ui/dropdown-menu';
   import * as Tabs from '$lib/components/ui/tabs';
-  import AvailableLoader from '$lib/features/drafts/draftees/available/loader.svelte';
-  import DraftedLoader from '$lib/features/drafts/draftees/drafted/loader.svelte';
-  import SystemLogsLoader from '$lib/features/drafts/system-logs/loader.svelte';
-  import { Button } from '$lib/components/ui/button';
   import type { Lab } from '$lib/features/drafts/types';
 
   import LabRoundSummary from './lab-round-summary.svelte';
+  import StudentsSummary from './students-summary.svelte';
+
+  import SystemLogsLoader from './system-logs/loader.svelte';
 
   interface Props {
     draftId: string;
@@ -24,7 +21,6 @@
   const { draftId, requestedAt, round, labs }: Props = $props();
 
   let group = $state<'students' | 'labs' | 'logs'>('students');
-  let selectedView = $state<'pending' | 'drafted'>('pending');
 </script>
 
 <Tabs.Root bind:value={group}>
@@ -46,45 +42,7 @@
   </div>
   <Tabs.Content value="students">
     {#if group === 'students'}
-      <div class="mb-2 flex justify-center xs:justify-start">
-        <DropdownMenu.Root>
-          <DropdownMenu.Trigger>
-            {#snippet child({ props })}
-              <Button
-                variant="outline"
-                class="bg-background hover:bg-accent dark:bg-input dark:hover:bg-input/80"
-                {...props}
-              >
-                <ChevronDownIcon class="size-4 text-muted-foreground transition-transform" />
-                {selectedView === 'pending' ? 'Pending Selection' : 'Already Drafted'}
-              </Button>
-            {/snippet}
-          </DropdownMenu.Trigger>
-          <DropdownMenu.Content align="start">
-            <DropdownMenu.Item onclick={() => (selectedView = 'pending')}>
-              <span>Pending Selection</span>
-            </DropdownMenu.Item>
-            <DropdownMenu.Item onclick={() => (selectedView = 'drafted')}>
-              <span>Already Drafted</span>
-            </DropdownMenu.Item>
-          </DropdownMenu.Content>
-        </DropdownMenu.Root>
-      </div>
-      {#if selectedView === 'pending'}
-        <span class="text-sm text-muted-foreground"
-          >Review undrafted students available for selection.</span
-        >
-        <div class="flex min-h-0 grow flex-col overflow-y-auto px-4 pb-4">
-          <AvailableLoader {draftId}>No available draftees.</AvailableLoader>
-        </div>
-      {:else if selectedView === 'drafted'}
-        <span class="text-sm text-muted-foreground"
-          >Review students who have already been assigned.</span
-        >
-        <div class="flex min-h-0 grow flex-col overflow-y-auto px-4 pb-4">
-          <DraftedLoader {draftId}>No drafted students yet.</DraftedLoader>
-        </div>
-      {/if}
+      <StudentsSummary {draftId} />
     {/if}
   </Tabs.Content>
   <Tabs.Content value="labs" class="min-w-0 overflow-auto">
